Flip back to usize.

+++ day3/src/main.rs
@@ -5,19 +5,19 @@ use std::io::{self, BufRead};
 
 const PRIORITIES: &str = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ";
 
-fn get_priority(c: char) -> u64 {
-    (PRIORITIES.find(c).unwrap_or(0) + 1) as u64
+fn get_priority(c: char) -> usize {
+    PRIORITIES.find(c).unwrap_or(0) + 1
 }
 
 fn problem1(file: File) -> io::Result<()> {
-    let mut total = 0u64;
+    let mut total = 0usize;
     for line in io::BufReader::new(file).lines().flatten() {
         total += line[..line.len() / 2]
             .chars()
             .unique()
             .filter(|x| line[line.len() / 2..].contains(*x))
             .map(get_priority)
-            .sum::<u64>();
+            .sum::<usize>();
     }
 
     println!("Problem 1 Score: {}", total);
@@ -25,7 +25,7 @@ fn problem1(file: File) -> io::Result<()> {
 }
 
 fn problem2(file: File) -> io::Result<()> {
-    let mut total = 0u64;
+    let mut total = 0usize;
 
     for group in io::BufReader::new(file)
         .lines()
@@ -39,7 +39,7 @@ fn problem2(file: File) -> io::Result<()> {
             .unique()
             .filter(|x| bags[1].contains(*x) && bags[2].contains(*x))
             .map(get_priority)
-            .sum::<u64>();
+            .sum::<usize>();
     }
 
     println!("Problem 2 Score: {}", total);
